IGNITE-17909 Extract Network configuration
