+++ modules/api/src/main/java/org/apache/ignite/IgnitionManager.java
@@ -48,8 +48,7 @@ public class IgnitionManager {
      * REST endpoint is functional).
      *
      * @param nodeName Name of the node. Must not be {@code null}.
-     * @param configStr Optional node configuration based on
-     *      {@link org.apache.ignite.configuration.schemas.network.NetworkConfigurationSchema}.
+     * @param configStr Optional node configuration.
      *      Following rules are used for applying the configuration properties:
      *      <ol>
      *        <li>Specified property overrides existing one or just applies itself if it wasn't
