+++ modules/network/src/main/resources/META-INF/services/org.apache.ignite.internal.configuration.ConfigurationModule
@@ -14,4 +14,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-org.apache.ignite.network.configuration.NetworkConfigurationModule
+org.apache.ignite.internal.network.configuration.NetworkConfigurationModule
