+++ modules/network/src/testFixtures/java/org/apache/ignite/network/scalecube/TestScaleCubeClusterServiceFactory.java
@@ -18,7 +18,7 @@
 package org.apache.ignite.network.scalecube;
 
 import io.scalecube.cluster.ClusterConfig;
-import org.apache.ignite.configuration.schemas.network.ClusterMembershipView;
+import org.apache.ignite.internal.network.configuration.ClusterMembershipView;
 
 /**
  * Scalecube test factory. Provides fast detection time.
